requirements.txt
accelerate==0.31.0
acres==0.3.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
aura-sr==0.0.4
backoff==2.2.1
blinker==1.9.0
branca==0.7.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.2.0
cryptography==43.0.1
ctransformers==0.2.27
dataclasses-json==0.6.7
decorator==5.1.1
diffusers==0.32.2
distro==1.9.0
einops==0.8.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.1
filelock==3.15.4
folium==0.17.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
geocoder==1.38.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langsmith==0.1.82
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==2.1.5
marshmallow==3.21.3
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.33.0
networkx==3.3
nibabel==5.3.2
nipype==1.10.0
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.70.0
opencage==3.0.2
openpyxl==3.1.2
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
phonenumbers==8.13.45
pillow==10.4.0
protobuf==5.29.4
prov==2.0.1
psutil==6.0.0
puremagic==1.28
py-cpuinfo==9.0.0
pyarrow==19.0.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydot==3.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2023.3.post1
pyxnat==1.6.3
PyYAML==6.0.1
ratelim==0.1.6
rdflib==6.3.2
referencing==0.36.2
regex==2024.5.15
requests==2.32.3
rpds-py==0.24.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.25.0
sentence-transformers==3.1.1
sentencepiece==0.2.0
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
streamlit==1.44.0
sympy==1.13.1
tabula-py==2.9.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.19.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.66.5
traits==7.0.2
transformers==4.42.3
trio==0.26.2
trio-websocket==0.11.1
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.9.4
zipp==3.21.0
